requirements.txt
discord.py>=2.3
httpx>=0.27
uvloop; platform_system!='Windows'
